+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/aip/reference/contributor/AipReferenceContributor.kt
@@ -48,8 +48,9 @@ class AipResourceReferenceProvider : PsiReferenceProvider() {
     private fun getReference(element: ProtobufStringValue): PsiReference? {
         val hover = element.parentOfType<ProtobufOptionHover>() ?: return null
         if (!hover.isOption(AipOptions.resourceReferenceOption)) return null
-        if (hover.value(AipOptions.resourceTypeField)?.stringValue != element) return null
-        return AipResourceReference(element)
+        if (hover.value(AipOptions.resourceTypeField)?.stringValue == element) return AipResourceReference(element)
+        if (hover.value(AipOptions.resourceChildTypeField)?.stringValue == element) return AipResourceReference(element)
+        return null
     }
 }