+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/completion/BuiltInTypeProvider.kt
@@ -1,8 +1,11 @@
 package io.kanro.idea.plugin.protobuf.lang.completion
 
+import com.intellij.codeInsight.completion.AddSpaceInsertHandler
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.lookup.LookupElement
+import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
 import io.kanro.idea.plugin.protobuf.lang.support.BuiltInType
 
@@ -18,6 +21,11 @@ class BuiltInTypeProvider : CompletionProvider<CompletionParameters>() {
     companion object {
         private val builtInTypes = BuiltInType.values().map {
             builtInTypeElement(it.value())
+        }
+
+        fun builtInTypeElement(keyword: String): LookupElement {
+            return LookupElementBuilder.create(keyword).withTypeText("built-in")
+                .withInsertHandler(AddSpaceInsertHandler.INSTANCE)
         }
     }
 }