+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/file/FileResolver.kt
@@ -1,12 +1,14 @@
 package io.kanro.idea.plugin.protobuf.lang.file
 
+import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.extensions.ExtensionPointName
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
-import io.kanro.idea.plugin.protobuf.lang.ProtobufFileType
+import kotlin.time.ExperimentalTime
+import kotlin.time.measureTimedValue
 
 interface FileResolver {
     companion object : FileResolver {
@@ -43,10 +45,15 @@ interface FileResolver {
             }
         }
 
+        @OptIn(ExperimentalTime::class)
         override fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile> {
-            return extensionPoint.extensionList.asSequence().flatMap {
-                it.collectProtobuf(path, module).asSequence()
-            }.asIterable()
+            val test = measureTimedValue {
+                extensionPoint.extensionList.asSequence().flatMap {
+                    it.collectProtobuf(path, module).asSequence()
+                }.asIterable()
+            }
+            Logger.getInstance(FileResolver::class.java).error("Collect protobuf in ${test.duration}")
+            return test.value
         }
 
         override fun collectProtobuf(path: String, project: Project): Iterable<VirtualFile> {
@@ -54,43 +61,13 @@ interface FileResolver {
                 it.collectProtobuf(path, project).asSequence()
             }.asIterable()
         }
-
-        fun collectProtobuf(
-            file: VirtualFile,
-            result: MutableList<VirtualFile> = mutableListOf()
-        ): Iterable<VirtualFile> {
-            if (!file.isDirectory) return result
-            for (child in file.children) {
-                if (child.isDirectory && checkDirectoryContainsProtobufFile(child)) {
-                    result += child
-                } else if (child.fileType is ProtobufFileType) {
-                    result += child
-                }
-            }
-            return result
-        }
-
-        private fun checkDirectoryContainsProtobufFile(file: VirtualFile): Boolean {
-            for (child in file.children) {
-                if (child.isDirectory) {
-                    return checkDirectoryContainsProtobufFile(child)
-                } else if (child.fileType is ProtobufFileType) {
-                    return true
-                }
-            }
-            return false
-        }
     }
 
     fun findFile(path: String, project: Project): Iterable<VirtualFile>
 
     fun findFile(path: String, module: Module): Iterable<VirtualFile>
 
-    fun collectProtobuf(path: String, module: Project): Iterable<VirtualFile> {
-        return listOf()
-    }
+    fun collectProtobuf(path: String, project: Project): Iterable<VirtualFile>
 
-    fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile> {
-        return listOf()
-    }
+    fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile>
 }