+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/file/ImportRootsFileResolver.kt
@@ -7,7 +7,7 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileManager
 import io.kanro.idea.plugin.protobuf.lang.settings.ProtobufSettings
 
-class ImportRootsFileResolver : FileResolver {
+class ImportRootsFileResolver : RootsFileResolver() {
     private fun getImportRoots(project: Project): List<VirtualFile> {
         val settings = ServiceManager.getService(project, ProtobufSettings::class.java)
         return settings.state.importRoots.mapNotNull {
@@ -15,27 +15,11 @@ class ImportRootsFileResolver : FileResolver {
         }
     }
 
-    override fun findFile(path: String, project: Project): Iterable<VirtualFile> {
-        return getImportRoots(project).mapNotNull {
-            it.findFileByRelativePath(path)?.takeIf { it.exists() }
-        }
+    override fun getRoots(project: Project): Iterable<VirtualFile> {
+        return getImportRoots(project)
     }
 
-    override fun findFile(path: String, module: Module): Iterable<VirtualFile> {
-        return findFile(path, module.project)
-    }
-
-    override fun collectProtobuf(path: String, project: Project): Iterable<VirtualFile> {
-        val result = mutableListOf<VirtualFile>()
-        getImportRoots(project).forEach {
-            val directory = it.findFileByRelativePath(path) ?: return@forEach
-            if (!directory.isDirectory) return@forEach
-            FileResolver.collectProtobuf(directory, result)
-        }
-        return result
-    }
-
-    override fun collectProtobuf(path: String, module: Module): Iterable<VirtualFile> {
-        return collectProtobuf(path, module.project)
+    override fun getRoots(module: Module): Iterable<VirtualFile> {
+        return getImportRoots(module.project)
     }
 }